Refactor Redux files to use original file structure, fix imports
