+++ app/src/actions/actionTypes.js
@@ -1,4 +1,6 @@
-// export const ACTIONTYPE = 'ACTION_TYPE';
-
 export const LOADING = 'LOADING';
 export const ERROR = 'ERROR';
+export const SEARCH_ITEMS = 'SEARCH_ITEMS';
+export const FILTER_ITEMS = 'FILTER_ITEMS';
+export const TOGGLE_WISHLIST = 'FILTER_ITEMS';
+export const REMOVE_WISH = 'REMOVE_WISH';
