+++ app/src/components/user/UserWrapper.jsx
@@ -26,7 +26,7 @@ class UserWrapper extends Component {
 
 const mapStateToProps = state => {
   return ({
-    user: state.userReducer,
+    user: state.user,
   });
 }
